--- tests/test_languages.py
@@ -1,40 +1,0 @@
-# Tests for language-specific stuff, often in response to issues.
-import sys
-from pathlib import Path
-
-from tested.configs import create_bundle
-from tested.datatypes import BasicBooleanTypes, BasicNumericTypes, BasicStringTypes
-from tested.languages.conventionalize import submission_name
-from tested.languages.generation import generate_statement
-from tested.serialisation import (
-    BooleanType,
-    FunctionCall,
-    FunctionType,
-    NumberType,
-    StringType,
-)
-from tested.testsuite import Suite
-from tests.manual_utils import configuration
-
-
-def test_function_arguments_without_brackets(tmp_path: Path, pytestconfig):
-    conf = configuration(pytestconfig, "", "haskell", tmp_path)
-    plan = Suite()
-    bundle = create_bundle(conf, sys.stdout, plan)
-
-    statement = FunctionCall(
-        type=FunctionType.FUNCTION,
-        name="test",
-        namespace=None,
-        arguments=[
-            NumberType(type=BasicNumericTypes.REAL, data=5.5),
-            StringType(type=BasicStringTypes.TEXT, data="hallo"),
-            BooleanType(type=BasicBooleanTypes.BOOLEAN, data=True),
-        ],
-    )
-
-    result = generate_statement(bundle, statement)
-    assert (
-        result
-        == f'{submission_name(bundle.lang_config)}.test 5.5 :: Double "hallo" True'
-    )
