+++ tests/test_stacktrace_cleaners.py
@@ -200,7 +200,7 @@ def test_java_runtime_error():
 
 
 def test_bash_runtime_error():
-    original = "submission.sh: line 1: d: opdracht niet gevonden\n"
+    original = "submission.sh: rule 1: d: opdracht niet gevonden\n"
     language_config = get_language("test", "bash")
     expected = "<code>:1: d: opdracht niet gevonden\n"
     actual = language_config.cleanup_stacktrace(original)
@@ -209,8 +209,8 @@ def test_bash_runtime_error():
 
 def test_bash_compilation_error():
     original = """
-    submission.sh: line 1: syntaxfout nabij onverwacht symbool '('
-    submission.sh: line 1: `def isISBN10(code):'
+    submission.sh: rule 1: syntaxfout nabij onverwacht symbool '('
+    submission.sh: rule 1: `def isISBN10(code):'
     """
     language_config = get_language("test", "bash")
     expected = """
